add POST /search-history-off api

+++ module/searchHistoryControl.js
@@ -9,24 +9,30 @@ const addSearchHistory = (searchKeyword = '', loginUserEmail = '') => {
         if(searchHistoryDataValidCheck(searchKeyword)){
             try{
                 await pgClient.connect();
-                
-                //SELECT search history
-                const selectHistorySql = 'SELECT * FROM shoot.search_history WHERE channel_email = $1 ORDER BY search_keyword_idx DESC';
-                const selectHistoryResult = await pgClient.query(selectHistorySql, [loginUserEmail]);
 
-                //DELETE duplicate history, last history
-                const keywordIndex = selectHistoryResult.rows.map(history => history.search_keyword).indexOf(searchKeyword);
-                if(keywordIndex !== -1){
-                    const deleteHistorySql = 'DELETE FROM shoot.search_history WHERE search_keyword_idx = $1';
-                    await pgClient.query(deleteHistorySql, [selectHistoryResult.rows[keywordIndex].search_keyword_idx]);
-                }else if(selectHistoryResult.rows.length === 5){
-                    const deleteHistorySql = 'DELETE FROM shoot.search_history WHERE search_keyword_idx = $1';
-                    await pgClient.query(deleteHistorySql, [selectHistoryResult.rows[4].search_keyword_idx]);
+                //SELECT search history off
+                const selectHistoryOffSql = 'SELECT * FROM shoot.search_history_off WHERE channel_email = $1';
+                const selectHistoryOffResult = await pgClient.query(selectHistoryOffSql, [loginUserEmail]);
+
+                if(selectHistoryOffResult.rows.length === 0){
+                    //SELECT search history
+                    const selectHistorySql = 'SELECT * FROM shoot.search_history WHERE channel_email = $1 ORDER BY search_keyword_idx DESC';
+                    const selectHistoryResult = await pgClient.query(selectHistorySql, [loginUserEmail]);
+
+                    //DELETE duplicate history, last history
+                    const keywordIndex = selectHistoryResult.rows.map(history => history.search_keyword).indexOf(searchKeyword);
+                    if(keywordIndex !== -1){
+                        const deleteHistorySql = 'DELETE FROM shoot.search_history WHERE search_keyword_idx = $1';
+                        await pgClient.query(deleteHistorySql, [selectHistoryResult.rows[keywordIndex].search_keyword_idx]);
+                    }else if(selectHistoryResult.rows.length === 5){
+                        const deleteHistorySql = 'DELETE FROM shoot.search_history WHERE search_keyword_idx = $1';
+                        await pgClient.query(deleteHistorySql, [selectHistoryResult.rows[4].search_keyword_idx]);
+                    }
+
+                    //INSERT
+                    const insertHistorySql = 'INSERT INTO shoot.search_history (search_keyword, channel_email) VALUES ($1, $2)';
+                    await pgClient.query(insertHistorySql, [searchKeyword, loginUserEmail]);
                 }
-
-                //INSERT
-                const insertHistorySql = 'INSERT INTO shoot.search_history (search_keyword, channel_email) VALUES ($1, $2)';
-                await pgClient.query(insertHistorySql, [searchKeyword, loginUserEmail]);
 
                 resolve(1);
             }catch(err){
@@ -45,6 +51,29 @@ const addSearchHistory = (searchKeyword = '', loginUserEmail = '') => {
     });
 }
 
+const getAllSearchHistory = (loginUserEmail = '', size = 5) => {
+    return new Promise(async (resolve, reject) => {
+        const pgClient = new Client(pgConfig);
+
+        try{
+            await pgClient.connect();
+            
+            //SELECT search history
+            const selectSearchHistorySql = 'SELECT search_keyword_idx, search_keyword AS keyword FROM shoot.search_keyword WHERE channel_email = $1';
+            const selectSearchHistoryResult = await pgClient.query(selectSearchHistorySql, [loginUserEmail]);
+
+            resolve(selectSearchHistoryResult.rows);
+        }catch(err){
+            reject({
+                statusCode: 409,
+                message : 'unexpected error occured',
+                err : err
+            });
+        }
+    });
+}
+
 module.exports = {
-    addSearchHistory : addSearchHistory
+    addSearchHistory : addSearchHistory,
+    getAllSearchHistory : getAllSearchHistory
 }
